Uploaded files
Traversal of a grid, walls unknown, in an attempt to find a red colored
node. After a section is traversed, the system calculates any sensed
walls and the all available open areas and goes to the next largest.
Once the red spot is found, the system calculates the fastest route back
home, going back a way it knows. The driver.c contains test methods,
driver methods, sorry it's in one file, that's the only way ceenbots
work. Two files are no go.